moved functions to view_funcs
